Rework merge logic
