+++ crates/nu-command/src/filters/merge/mod.rs
@@ -1,0 +1,4 @@
+mod common;
+pub mod merge_;
+
+pub use merge_::Merge;
